requirements.txt
gunicorn == 20.0.4
flask == 2.0.1
requests == 2.26.0
beautifulsoup4 ==4.8.2
Werkzeug==2.2.2

pandas
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.26
langchain-openai==0.1.8
python-dotenv==1.0.1
langchainhub
langsmith